Simplify the ai-tools CLI (#58066)
